+++ changes.md
@@ -2,6 +2,8 @@
 
 ## 1.3.1 - Unreleased
 
+- Collapsible pattern: add a new `slide-horizontal` transition.
+
 - Slideshow builder pattern: make it possible to insert the slideshow fieldset
   in a different location than at the start of the form.